+++ src/Jimu.Server/Auth/Middlewares/JwtAuthorizationOptions.cs
@@ -34,7 +34,7 @@ namespace Jimu.Server.Auth.Middlewares
             var payload = new Dictionary<string, object>();
             if (ValidateLifetime)
             {
-                payload["exp"] = DateTime.Now.AddMinutes(ExpireTimeSpan.TotalMinutes).ToInt();
+                payload["exp"] = DateTime.Now.AddMinutes(ExpireTimeSpan.TotalMinutes).ToLong();
             }
             if (ValidateIssuer)
             {
